requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict==2.4.4
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==5.3.0
msgpack==1.1.0
multitasking==0.0.11
numpy==2.1.1
oauthlib==3.2.2
pandas==2.2.3
peewee==3.17.6
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.43
zope.interface==7.0.3
